requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
celery==5.4.0
redis==5.0.4
pydantic==2.7.0
pydantic-settings==2.2.1
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
openai-whisper==20231117
torch==2.3.0
google-generativeai
chromadb
sentence-transformers==3.0.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
requests
groq
